+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Does not report screen events if there's no title ([#32](https://github.com/PostHog/posthog-android/pull/32))
 - Add `distinctId()` getter to the Public API ([#33](https://github.com/PostHog/posthog-android/pull/33))
+- Add compatibility to Java 8 bytecode (previously Java 11+) ([#34](https://github.com/PostHog/posthog-android/pull/34))
 
 ## 3.0.0-alpha.4 - 2023-10-03