+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Next
+## 3.0.0-alpha.5 - 2023-10-04
 
 - Does not report screen events if there's no title ([#32](https://github.com/PostHog/posthog-android/pull/32))
 - Add `distinctId()` getter to the Public API ([#33](https://github.com/PostHog/posthog-android/pull/33))